Fix/refactor dynamo ipex backend
[ghstack-poisoned]
