+++ benchmarks/dynamo/common.py
@@ -2073,18 +2073,7 @@ def run(runner, args, original_dir=None):
         optimize_ctx = nothing
         output_filename = "nothing.csv"
     elif args.backend:
-        if args.backend == "ipex":
-            if args.bfloat16:
-                optimize_ctx = torch._dynamo.optimize(
-                    backends.ipex_bf16, nopython=args.nopython
-                )
-            else:
-                assert args.float32, "IPEX only supports fp32 and bf16 for now."
-                optimize_ctx = torch._dynamo.optimize(
-                    backends.ipex_fp32, nopython=args.nopython
-                )
-        else:
-            optimize_ctx = torch._dynamo.optimize(args.backend, nopython=args.nopython)
+        optimize_ctx = torch._dynamo.optimize(args.backend, nopython=args.nopython)
         experiment = speedup_experiment
         if args.accuracy:
             output_filename = f"accuracy_{args.backend}.csv"
