+++ test/dynamo/test_optimizations.py
@@ -1,29 +1,14 @@
 # Owner(s): ["module: dynamo"]
-import importlib
 import unittest
 
 import torch
 
 import torch._dynamo
+import torch._dynamo.backends.ipex
 import torch._dynamo.test_case
-from torch._dynamo.optimizations import backends
+from torch._dynamo.backends.ipex import has_ipex
+from torch._dynamo.backends.onnxrt import has_onnxruntime
 from torch._dynamo.testing import same
-
-
-def has_onnxruntime():
-    try:
-        importlib.import_module("onnxruntime")
-        return True
-    except ImportError:
-        return False
-
-
-def has_ipex():
-    try:
-        importlib.import_module("intel_extension_for_pytorch")
-        return True
-    except ImportError:
-        return False
 
 
 class Seq(torch.nn.Module):
@@ -114,7 +99,7 @@ class TestOptimizations(torch._dynamo.test_case.TestCase):
         model = model.eval()
         input = torch.randn(8, 3, 64, 64).contiguous(memory_format=torch.channels_last)
         r1 = model(input)
-        opt_model = torch._dynamo.optimize(backends.ipex_fp32)(model)
+        opt_model = torch._dynamo.optimize(torch._dynamo.backends.ipex.ipex_fp32)(model)
         with torch.no_grad():
             r2 = opt_model(input)
         self.assertTrue(same(r1, r2))
@@ -127,7 +112,7 @@ class TestOptimizations(torch._dynamo.test_case.TestCase):
         model = model.eval()
         input = torch.randn(8, 3, 64, 64).contiguous(memory_format=torch.channels_last)
         r1 = model(input)
-        opt_model = torch._dynamo.optimize(backends.ipex_bf16)(model)
+        opt_model = torch._dynamo.optimize(torch._dynamo.backends.ipex.ipex_bf16)(model)
         with torch.no_grad(), torch.cpu.amp.autocast():
             r2 = opt_model(input)
         self.assertTrue(same(r1, r2.float(), tol=0.1))
