+++ test/dynamo/test_verify_correctness.py
@@ -1,31 +1,15 @@
 # Owner(s): ["module: dynamo"]
-import importlib
 import operator
 import unittest
 
 import torch
 
 import torch._dynamo
+import torch._dynamo.backends.ipex
 import torch._dynamo.config as config
 import torch._dynamo.test_case
-from torch._dynamo.optimizations import backends
+from torch._dynamo.backends.ipex import has_ipex
 from torch._dynamo.testing import same
-
-
-def has_onnxruntime():
-    try:
-        importlib.import_module("onnxruntime")
-        return True
-    except ImportError:
-        return False
-
-
-def has_ipex():
-    try:
-        importlib.import_module("intel_extension_for_pytorch")
-        return True
-    except ImportError:
-        return False
 
 
 class Seq(torch.nn.Module):
@@ -161,7 +145,7 @@ class TestVerifyCorrectness(torch._dynamo.test_case.TestCase):
         model = model.eval()
         input = torch.randn(8, 3, 64, 64).contiguous(memory_format=torch.channels_last)
         r1 = model(input)
-        opt_model = torch._dynamo.optimize(backends.ipex_fp32)(model)
+        opt_model = torch._dynamo.optimize(torch._dynamo.backends.ipex.ipex_fp32)(model)
         with torch.no_grad():
             r2 = opt_model(input)
         self.assertTrue(same(r1, r2))
