+++ torch/_dynamo/backends/common.py
@@ -107,3 +107,9 @@ def device_from_inputs(example_inputs) -> torch.device:
     for x in example_inputs:
         if hasattr(x, "device"):
             return x.device
+
+
+def dtype_from_inputs(example_inputs) -> torch.dtype:
+    for x in example_inputs:
+        if hasattr(x, "dtype"):
+            return x.dtype
